chore(zendesk-ticket): rename folder and files
