+++ packages/cli-zendesk-ticket/src/index.ts
@@ -1,5 +1,5 @@
-import Server from './Server'
+import Cli from './cli'
 import dotenv from 'dotenv'
 dotenv.config()
-const app = new Server()
+const app = new Cli()
 app.start()
